% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaDelta-methods-beta-delta.R
\name{coef.betadelta}
\alias{coef.betadelta}
\title{Standardized Regression Slopes}
\usage{
\method{coef}{betadelta}(object, ...)
}
\arguments{
\item{object}{Object of class \code{betadelta}.}

\item{...}{additional arguments.}
}
\value{
Returns a vector of standardized regression slopes.
}
\description{
Standardized Regression Slopes
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaDelta(object)
coef(std)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
