% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaMethod-methods.R
\name{coef.deltamethod}
\alias{coef.deltamethod}
\title{Estimates}
\usage{
\method{coef}{deltamethod}(object, ...)
}
\arguments{
\item{object}{Object of class \code{deltamethod}.}

\item{...}{additional arguments.}
}
\value{
Returns a vector of estimates.
}
\description{
Estimates
}
\examples{
object <- glm(
  formula = vs ~ wt + disp,
  family = "binomial",
  data = mtcars
)
def <- list("exp(wt)", "exp(disp)")
out <- DeltaGeneric(
  object = object,
  def = def,
  alpha = 0.05
)
coef(out)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
