% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaDelta-methods-diff-beta-delta.R
\name{coef.diffbetadelta}
\alias{coef.diffbetadelta}
\title{Differences of Standardized Regression Slopes}
\usage{
\method{coef}{diffbetadelta}(object, ...)
}
\arguments{
\item{object}{Object of class \code{diffbetadelta}.}

\item{...}{additional arguments.}
}
\value{
Returns a vector of differences of standardized regression slopes.
}
\description{
Differences of Standardized Regression Slopes
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaDelta(object)
diff <- DiffBetaDelta(std)
coef(diff)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
