% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaDelta-methods-r-sq-delta.R
\name{coef.rsqbetadelta}
\alias{coef.rsqbetadelta}
\title{R-Squared and Adjusted R-Squared}
\usage{
\method{coef}{rsqbetadelta}(object, ...)
}
\arguments{
\item{object}{Object of class \code{rsqrdelta}.}

\item{...}{additional arguments.}
}
\value{
Returns a vector of R-squared and adjusted R-squared.
}
\description{
R-Squared and Adjusted R-Squared
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaDelta(object)
out <- rsq(std)
coef(out)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
