% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaDelta-methods-r-sq-delta.R
\name{confint.rsqbetadelta}
\alias{confint.rsqbetadelta}
\title{Confidence Intervals for R-Squared and Adjusted R-Squared}
\usage{
\method{confint}{rsqbetadelta}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{Object of class \code{rsqrdelta}.}

\item{parm}{a specification of which parameters
are to be given confidence intervals,
either a vector of numbers or a vector of names.
If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of confidence intervals.
}
\description{
Confidence Intervals for R-Squared and Adjusted R-Squared
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaDelta(object)
out <- rsq(std)
confint(out, level = 0.95)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
