% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaDelta-diff-beta-delta.R
\name{dif.betadelta}
\alias{dif.betadelta}
\title{Differences of Standardized Regression Slopes}
\usage{
\method{dif}{betadelta}(object)
}
\arguments{
\item{object}{Object of class \code{betadelta}.}
}
\value{
Returns an object of class \code{betadelta}
which is a list with the following elements:
\describe{
\item{call}{Function call.}
\item{betadelta}{Object of class \code{betadelta}.}
\item{est}{Vector of differences of standardized regression slopes.}
\item{vcov}{Sampling covariance matrix
of differences of standardized slopes.}
\item{acov}{Asymptotic covariance matrix
of differences of standardized slopes.}
}
}
\description{
Differences of Standardized Regression Slopes
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaDelta(object)
out <- dif(std)
# Methods -------------------------------------------------------
print(out)
summary(out)
coef(out)
vcov(out)
confint(out, level = 0.95)
}
\seealso{
Other Beta Delta Functions: 
\code{\link{BetaDelta}()},
\code{\link{dif}()},
\code{\link{rsq.betadelta}()},
\code{\link{rsq}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Beta Delta Functions}
\keyword{betaDelta}
\keyword{dif}
\keyword{methods}
