% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaDelta-generic-diff-beta-delta.R
\name{dif}
\alias{dif}
\title{Differences of Standardized Regression Slopes}
\usage{
dif(object)
}
\arguments{
\item{object}{Object used to select a method.}
}
\description{
Differences of Standardized Regression Slopes
}
\seealso{
Other Beta Delta Functions: 
\code{\link{BetaDelta}()},
\code{\link{dif.betadelta}()},
\code{\link{rsq.betadelta}()},
\code{\link{rsq}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Beta Delta Functions}
\keyword{betaDelta}
\keyword{dif}
\keyword{methods}
