% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaDelta-methods-diff-beta-delta.R
\name{print.difbetadelta}
\alias{print.difbetadelta}
\title{Print Method for an Object of Class \code{difbetadelta}}
\usage{
\method{print}{difbetadelta}(x, alpha = c(0.05, 0.01, 0.001), digits = 4, ...)
}
\arguments{
\item{x}{Object of class \code{difbetadelta}.}

\item{alpha}{Significance level.}

\item{digits}{Digits to print.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of standardized regression slopes,
standard errors, test statistics, p-values, and confidence intervals.
}
\description{
Print Method for an Object of Class \code{difbetadelta}
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaDelta(object)
out <- dif(std)
print(out)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
