% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaDelta-r-sq-delta.R
\name{rsq.betadelta}
\alias{rsq.betadelta}
\title{Estimate Coefficient of Determination
(R-Squared and Adjusted R-Squared)
and Sampling Covariance Matrix}
\usage{
\method{rsq}{betadelta}(object)
}
\arguments{
\item{object}{Object of class \code{betadelta}.}
}
\value{
Returns an object of class \code{rsqbetadelta}
which is a list with the following elements:
\describe{
\item{betadelta}{Object of class \code{betadelta}.}
\item{est}{R-squared (\eqn{R^{2}}) and
adjusted R-squared (\eqn{\bar{R}^{2}}).}
\item{vcov}{Sampling covariance matrix of R-squared (\eqn{R^{2}}) and
adjusted R-squared (\eqn{\bar{R}^{2}}).}
\item{acov}{Asymptotic covariance matrix of R-squared (\eqn{R^{2}}) and
adjusted R-squared (\eqn{\bar{R}^{2}}).}
}
}
\description{
Estimate Coefficient of Determination
(R-Squared and Adjusted R-Squared)
and Sampling Covariance Matrix
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaDelta(object)
out <- rsq(std)
# Methods -------------------------------------------------------
print(out)
summary(out)
coef(out)
vcov(out)
confint(out, level = 0.95)
}
\seealso{
Other Beta Delta Functions: 
\code{\link{BetaDelta}()},
\code{\link{dif.betadelta}()},
\code{\link{dif}()},
\code{\link{rsq}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Beta Delta Functions}
\keyword{betaDelta}
\keyword{methods}
\keyword{rsq}
