% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaDelta-generic-r-sq-delta.R
\name{rsq}
\alias{rsq}
\title{Estimate Coefficient of Determination
(R-Squared and Adjusted R-Squared)
and Sampling Covariance Matrix}
\usage{
rsq(object)
}
\arguments{
\item{object}{Object used to select a method.}
}
\description{
Estimate Coefficient of Determination
(R-Squared and Adjusted R-Squared)
and Sampling Covariance Matrix
}
\seealso{
Other Beta Delta Functions: 
\code{\link{BetaDelta}()},
\code{\link{dif.betadelta}()},
\code{\link{dif}()},
\code{\link{rsq.betadelta}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Beta Delta Functions}
\keyword{betaDelta}
\keyword{methods}
\keyword{rsq}
