% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaDelta-methods-beta-delta.R
\name{vcov.betadelta}
\alias{vcov.betadelta}
\title{Sampling Covariance Matrix of the Standardized Regression Slopes}
\usage{
\method{vcov}{betadelta}(object, ...)
}
\arguments{
\item{object}{Object of class \code{betadelta}.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of the
variance-covariance matrix
of standardized slopes.
}
\description{
Sampling Covariance Matrix of the Standardized Regression Slopes
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaDelta(object)
vcov(std)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
