% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaDelta-methods-r-sq-delta.R
\name{vcov.rsqbetadelta}
\alias{vcov.rsqbetadelta}
\title{Sampling Covariance Matrix of R-Squared and Adjusted R-Squared}
\usage{
\method{vcov}{rsqbetadelta}(object, ...)
}
\arguments{
\item{object}{Object of class \code{rsqbetadelta}.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of the variance-covariance matrix
of R-squared and adjusted R-squared.
}
\description{
Sampling Covariance Matrix of R-Squared and Adjusted R-Squared
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaDelta(object)
out <- rsq(std)
vcov(out)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
