% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaDelta-package.R
\docType{package}
\name{betaDelta-package}
\alias{betaDelta-package}
\alias{_PACKAGE}
\title{betaDelta: Confidence Intervals for Standardized Regression Coefficients}
\description{
Generates confidence intervals for standardized regression coefficients using delta method standard errors for models fitted by lm() as described in Yuan and Chan (2011) \doi{10.1007/s11336-011-9224-6} and Jones and Waller (2015) \doi{10.1007/s11336-013-9380-y}. A description of the package and code examples are presented in Pesigan, Sun, and Cheung (2023) \doi{10.1080/00273171.2023.2201277}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeksterslab/betaDelta}
  \item \url{https://jeksterslab.github.io/betaDelta/}
  \item Report bugs at \url{https://github.com/jeksterslab/betaDelta/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

}
\keyword{internal}
