% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaMethod-delta-generic.R
\name{DeltaGeneric}
\alias{DeltaGeneric}
\title{Delta Method Confidence Intervals (Generic)}
\usage{
DeltaGeneric(
  func,
  mu,
  sigmacap,
  n,
  theta = 0,
  alpha = c(0.05, 0.01, 0.001),
  z = TRUE,
  df,
  test = FALSE
)
}
\arguments{
\item{func}{An R function the input of which is \code{mu}.
The function should return a vector of any length.}

\item{mu}{Numeric vector.
Input of \code{func}.}

\item{sigmacap}{Numeric vector or matrix.
Asymptotic covariance matrix of \code{mu}.}

\item{n}{Sample size.}

\item{theta}{Numeric vector.
Parameter values when the null hypothesis is true.}

\item{alpha}{Numeric vector.
Significance level/s.}

\item{z}{Logical.
If \code{z = TRUE},
use the standard normal distribution.
If \code{z = FALSE},
use the t distribution.}

\item{df}{Numeric.
Degrees of freedom if \code{z = FALSE}.}

\item{test}{Logical.
If \code{TRUE},
return only the results of hypothesis tests.
If \code{FALSE},
return both results of hypothesis tests and confidence intervals.}
}
\value{
Returns a numeric matrix with the following variables:
\describe{
\item{est}{Estimates}
\item{se}{Standard errors}
\item{t or z}{Test statistics}
\item{p}{p value}
\item{ci}{Confidence intervals}
}
Note that if \code{test = TRUE}, the \code{ci} columns are omitted.
}
\description{
Calculates delta method confidence intervals
for a function of parameters
using a numerical Jacobian.
}
\examples{
g <- function(x) {
  1 / x
}
mu <- 100
sigmasq <- 225
n <- 30
DeltaGeneric(
  func = g,
  mu = mu,
  sigmacap = sigmasq,
  n = n,
  alpha = 0.05
)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Delta Method Functions}
\keyword{deltaMethod}
